Introduce job utils
